requirements.txt
# LiveKit Agents with all plugins
livekit-agents[deepgram,openai,cartesia,silero,turn-detector,google,azure]~=1.2

# Noise cancellation plugin
livekit-plugins-noise-cancellation~=0.2

# Environment variables
python-dotenv

# HTTP requests for API calls
requests